Add HP alert for rare Diana mobs and refactor events
Introduces an HP alert feature for rare Diana mobs, sending a title alert when their health drops below a configurable threshold. Refactors PartyFinder event classes into a dedicated 'partyfinder' package and updates relevant imports. Also simplifies Mayor date calculation logic.
