+++ src/main/kotlin/net/sbo/mod/guis/Guis.kt
@@ -7,7 +7,7 @@ import net.sbo.mod.guis.partyfinder.PartyFinderGUI
 import net.sbo.mod.utils.chat.Chat
 import net.sbo.mod.utils.events.SBOEvent
 import net.sbo.mod.utils.events.Register
-import net.sbo.mod.utils.events.impl.PartyFinderOpenEvent
+import net.sbo.mod.utils.events.impl.partyfinder.PartyFinderOpenEvent
 import net.sbo.mod.utils.game.World
 import net.sbo.mod.utils.http.Http
