+++ src/main/kotlin/net/sbo/mod/partyfinder/PartyFinderManager.kt
@@ -25,7 +25,7 @@ import net.sbo.mod.utils.data.PartyUpdateResponse
 import net.sbo.mod.utils.data.Reqs
 import net.sbo.mod.utils.events.annotations.SboEvent
 import net.sbo.mod.utils.events.impl.game.DisconnectEvent
-import net.sbo.mod.utils.events.impl.PartyFinderRefreshListEvent
+import net.sbo.mod.utils.events.impl.partyfinder.PartyFinderRefreshListEvent
 import net.sbo.mod.utils.events.impl.game.ChatMessageEvent
 import net.sbo.mod.utils.http.Http.getInt
 import java.util.UUID
