+++ src/main/kotlin/net/sbo/mod/utils/events/impl/partyfinder/PartyFinderOpenEvent.kt
@@ -1,4 +1,4 @@
-package net.sbo.mod.utils.events.impl
+package net.sbo.mod.utils.events.impl.partyfinder
 
 /**
  * Called when the party finder GUI is opened.
