+++ src/main/kotlin/net/sbo/mod/utils/events/impl/partyfinder/PartyFinderRefreshListEvent.kt
@@ -1,4 +1,4 @@
-package net.sbo.mod.utils.events.impl
+package net.sbo.mod.utils.events.impl.partyfinder
 
 /**
  * Called when the party finder list is refreshed.
